move console api to it's own subfolder
